feat(configs): created configs for db and env
